+++ packages/models/src/Domain/Runtime/Deltas/OutOfSync.ts
@@ -8,9 +8,9 @@ import { extendArray } from '@standardnotes/utils'
 import { ImmutablePayloadCollection } from '../Collection/Payload/ImmutablePayloadCollection'
 import { isDecryptedPayload } from '../../Abstract/Payload/Interfaces/TypeCheck'
 import { PayloadContentsEqual } from '../../Utilities/Payload/PayloadContentsEqual'
-import { PayloadsByDuplicating } from '../../Utilities/Payload/PayloadsByDuplicating'
 import { PayloadsDelta } from './Delta'
 import { PayloadSource } from '../../Abstract/Payload/Types/PayloadSource'
+import { ConflictDelta } from './Conflict'
 
 type Return = EncryptedPayloadInterface | DecryptedPayloadInterface | DeletedPayloadInterface
 
@@ -22,37 +22,33 @@ export class DeltaOutOfSync extends PayloadsDelta<
   public async resultingCollection(): Promise<ImmutablePayloadCollection<Return>> {
     const results: Return[] = []
 
-    for (const payload of this.applyCollection.all()) {
-      /**
-       * Map the server payload as authoritive content. If client copy differs,
-       * we will create a duplicate of it below.
-       * This is also neccessary to map the updated_at value from the server
-       */
-      results.push(payload)
-
-      const base = this.findBasePayload(payload.uuid)
+    for (const apply of this.applyCollection.all()) {
+      const base = this.findBasePayload(apply.uuid)
       if (!base) {
+        results.push(apply)
         continue
       }
 
       const isBaseDecrypted = isDecryptedPayload(base)
-      const isApplyDecrypted = isDecryptedPayload(payload)
+      const isApplyDecrypted = isDecryptedPayload(apply)
+
       const needsConflict =
         isApplyDecrypted !== isBaseDecrypted ||
-        (isApplyDecrypted && isBaseDecrypted && !PayloadContentsEqual(payload, base))
+        (isApplyDecrypted && isBaseDecrypted && !PayloadContentsEqual(apply, base))
 
       if (needsConflict) {
-        /**
-         * We create a copy of the local existing item and sync that up.
-         * It will be a 'conflict' of itself
-         */
-        const copyResults = await PayloadsByDuplicating({
-          payload: base,
-          baseCollection: this.baseCollection,
-          isConflict: true,
-        })
-        extendArray(results, copyResults)
-        continue
+        const delta = new ConflictDelta(
+          this.baseCollection,
+          base,
+          apply,
+          PayloadSource.ConflictData,
+        )
+
+        const deltaCollection = await delta.resultingCollection()
+        const conflictResults = deltaCollection.all()
+        extendArray(results, conflictResults)
+      } else {
+        results.push(apply)
       }
     }